+++ CHANGELOG.md
@@ -5,3 +5,16 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 - Initial release
+
+## 0.0.4 - 2018-03-22
+
+Added a "build command from parameters" command. You can use a separate template file containing rules on how to format
+parameters into a command line (or similar).
+
+### Added
+- added build command
+- extension remembers used files until editor's quit
+- replacements and commands open in another column
+
+### Fixed
+- extension initializes on other commands